+++ index.ts
@@ -7,18 +7,18 @@ dotenv.config()
 const client = new DiscordJS.Client({
     intents: [
         Intents.FLAGS.GUILDS,
-        Intents.FLAGS.GUILD_MESSAGES
+        Intents.FLAGS.GUILD_MESSAGES,
+        Intents.FLAGS.GUILD_MESSAGE_REACTIONS,
     ]
 })
 
-client.on('ready', () => {
-    console.log(`the bot is ready`)
 
+client.on('ready', () => {
     new WOKCommands(client, {
         commandDir: path.join(__dirname, 'commands'),
         typeScript: true,
         testServers: ['763349277493559296'],
     })
+    .setBotOwner(['460466062593622016'])
 })
-
-client.login(process.env.TOKEN)
+client.login(process.env.TOKEN);